feat: Allow separate navigation, page interaction and saving snapshots
If you need to create multiple snapshots for a more complicated scenario
on an SPA, you will want to stay on the same page and do a couple of
actions. The objects in the `pages` array need only one of "url", "file"
and "wait" parameters to allow creating a sequence of commands.