fix: Report unknown instructions as errors
BREAKING CHANGE: If you include a key in your command object, which is not recognized as an instruction or other known key (file, options), the task execution will fail.

Earlier you could put any key to a command object. You could use it as a comment or to store any other data there. Unfortunately, it lead to mistakes like typos in instruction names, which were  difficult to detect. From now on, only recognized keys  are allowed to appear in commands. You can use JavaScript comments to explain what a particular command or instruction does in your scenario.