fix: Check existence earlier, than other characteristics like focused
BREAKING CHANGE: If more "is*" instructions are used, the order has been changed to prevent unnecessary blocks. Earlier, visibility was checked before other features, which could confuse the error reporting. The new order of evaluation of "is*" instruction reflects their importance. The instructions can be put to a single block, evaluated in the "natural" order and fail with the most important condition:

 * isExisting
 * isVisible
 * isVisibleWithinViewport
 * isEnabled
 * isSelected
 * isFocused
 * isNotExisting
 * isNotVisible
 * isNotVisibleWithinViewport
 * isNotEnabled
 * isNotFocused

This is very unlikely to break anything in real-world commands, but still, better to be sure.